feat!: drop stack in favor of native stack
