+++ docs/docs/features/options.md
@@ -3,7 +3,7 @@ title: Screen Options
 sidebar_position: 5
 ---
 
-Qualified layouts, like the ones found in `expo-router` (Stack, Tabs, NativeStack, Drawer, Layout) have a static `Screen` component which can be used to configure the behavior of a route declaratively.
+Qualified layouts, like the ones found in `expo-router` (Stack, Tabs, Drawer, Layout) have a static `Screen` component which can be used to configure the behavior of a route declaratively.
 
 All `Screen` components are the same and render `null`, but they have different types for convenience. The common props are:
