+++ packages/expo-router/src/index.tsx
@@ -5,7 +5,6 @@ export { useFocusEffect } from "@react-navigation/native";
 export { ErrorBoundaryProps } from "./views/Try";
 
 export { Stack } from "./layouts/Stack";
-export { NativeStack } from "./layouts/NativeStack";
 export { Tabs } from "./layouts/Tabs";
 export { Drawer } from "./layouts/Drawer";
 export { withLayoutContext } from "./layouts/withLayoutContext";
